perf(Lotto) : 로또 구매 금액 범위 확장

+++ src/main/java/lotto/controller/LottoPurchaseController.java
@@ -10,7 +10,7 @@ import java.util.List;
 public class LottoPurchaseController {
     // 로또 구매를 중계할 클래스입니다.
     public List<Lotto> purchaseLotto() {
-        int numberOfIssues = numberOfIssues();
+        long numberOfIssues = numberOfIssues();
         PurchasedLottoesView purchasedLottoesView = new PurchasedLottoesView();
         purchasedLottoesView.printNumberOfIssues(numberOfIssues);
 
@@ -20,13 +20,13 @@ public class LottoPurchaseController {
         return purchasedLotto;
     }
 
-    private int numberOfIssues() {
+    private long numberOfIssues() {
         LottoStoreCounter lottoStoreCounter = new LottoStoreCounter();
 
         return lottoStoreCounter.counter();
     }
 
-    private List<Lotto> purchasedLotto(int numberOfIssues) {
+    private List<Lotto> purchasedLotto(long numberOfIssues) {
         LottoStoreMachine lottoStoreMachine = new LottoStoreMachine();
 
         return lottoStoreMachine.saleLotto(numberOfIssues);
